Work around SQLAlchemy not being able to create default date
There only seem to exist ways to create default datetimes as strings and
when parsing those as a date, an error will occur (see
https://github.com/sqlalchemy/sqlalchemy/issues/10196).
Thus, we now create a datetime object directly when creating a new
entry (which can be converted to a date without issues)

Note that the fact that this default value will be computed by the
client is not a problem, since the possible delays shouldn't affect the
resulting dates (only the specific times within that date). And even if
they do, it's not that bad.