Make SSH tunneling work on Windows
By not explicitly specifying a local bind address when creating the SSH
tunnel, it would bind to 0.0.0.0 (all interfaces). However, for some
reason this was causing issues when trying to use Memmer on a Windows
host.

Since for our use-case we actually really want to restrict access to our
tunnel forwarder to connections from the same host machine, explicitly
binding to 127.0.0.1 is the better thing to do. Incidentally, this also
fixes the issue seen on Windows.